Validaciones completas

+++ NodeJs/src/Views/CompradorApp.js
@@ -10,9 +10,24 @@ export default function CompradorApp() {
   const tableRef = useRef(null);
   const [selectNameComprador, setSelectNameComprador] = useState(null);
   const listarComprador = () => {
+
+    //FORMATO EN TABLA 
+    const capitalize = (str) => {
+      return str.replace(/\w\S*/g, (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase());
+    };
+
     CompradorService.findAll()
       .then((response) => {
-        setComprador(response.data);
+        const formatoData = response.data.map(item => ({
+          ...item,
+          nombre: capitalize(item.nombre),
+          apellido_Pat: capitalize(item.apellido_Pat),
+          apellido_Mat: capitalize(item.apellido_Mat),
+          telefono: capitalize(item.telefono),
+          nombre_Empresa: capitalize(item.nombre_Empresa)
+      }));
+        setComprador(formatoData);
+        console.log(formatoData);
       })
       .catch((error) => {
         console.log(error);
@@ -41,7 +56,9 @@ export default function CompradorApp() {
   }, []);
 
   const deleteComprador = (id) => {
-    CompradorService.delete(id)
+    const confirmarDelete = window.confirm("¿Estás seguro de que deseas eliminar este registro?");
+    if (confirmarDelete) {
+      CompradorService.delete(id)
       .then((response) => {
         listarComprador();
         console.log(response.id);
@@ -49,6 +66,8 @@ export default function CompradorApp() {
       .catch((error) => {
         console.log(error);
       });
+    }
+   
   };
 
   const handleRowClick = (id, name) => {
@@ -68,41 +87,41 @@ export default function CompradorApp() {
             : ""}
         </p>
         <div className="table-container" ref={tableRef}>
-        <div className="table-responsive">
-        <table className="table table-hover table-bordered">
-            <thead className="table-success">
-              <tr>
-                <th>Id Comprador</th>
-                <th>Nombre del comprador</th>
-                <th>Telefono</th>
-                <th>Nombre de la empresa</th>
-              </tr>
-            </thead>
-            <tbody>
-              {comprador.map((comprador) => (
-                <tr
-                  key={comprador.id_Comprador}
-                  onClick={() =>
-                    handleRowClick(
-                      comprador.id_Comprador,
-                      comprador.nombre + " " + comprador.apellido_Pat
-                    )
-                  }
-                  className={
-                    selectedCompradorId === comprador.id_Comprador ? "selected" : ""
-                  }
-                >
-                  <td>{comprador.id_Comprador}</td>
-                  <td>
-                    {comprador.nombre} {comprador.apellido_Pat}{" "}
-                    {comprador.apellido_Mat}
-                  </td>
-                  <td>{comprador.telefono}</td>
-                  <td>{comprador.nombre_Empresa}</td>
+          <div className="table-responsive">
+            <table className="table table-hover table-bordered">
+              <thead className="table-success">
+                <tr>
+                  <th>Id Comprador</th>
+                  <th>Nombre del comprador</th>
+                  <th>Telefono</th>
+                  <th>Nombre de la empresa</th>
                 </tr>
-              ))}
-            </tbody>
-          </table>
+              </thead>
+              <tbody>
+                {comprador.map((comprador) => (
+                  <tr
+                    key={comprador.id_Comprador}
+                    onClick={() =>
+                      handleRowClick(
+                        comprador.id_Comprador,
+                        comprador.nombre + " " + comprador.apellido_Pat
+                      )
+                    }
+                    className={
+                      selectedCompradorId === comprador.id_Comprador ? "selected" : ""
+                    }
+                  >
+                    <td>{comprador.id_Comprador}</td>
+                    <td>
+                      {comprador.nombre} {comprador.apellido_Pat}{" "}
+                      {comprador.apellido_Mat}
+                    </td>
+                    <td>{comprador.telefono}</td>
+                    <td>{comprador.nombre_Empresa}</td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
           </div>
         </div>
         <div className="buttonsInLine">
@@ -123,7 +142,7 @@ export default function CompradorApp() {
                   Insertar
                 </button>
               </Link>
-             
+
             </>
           )}
           {rowSelected && (
